test(tcp): consolidate TCP test files into single tcp_test.go
Merged tcp_delta_test.go and tcp_collect_with_delta_test.go into the main
tcp_test.go file for better organization and maintainability.

Changes:
- Consolidated all TCP test functionality into tcp_test.go
- Fixed package qualification issues for collectors_test package
- Adapted private method tests to use public interface methods
- All tests continue to pass with the unified structure
